brew_requirements.txt
cppzmq
nlohmann-json
python@3.11
zeromq
